refactor: extract general string utils out of Validator/
XmlReader (runtime) now reuses TrimAsciiWhitespace, which lived under
Validator/Utils/ — a layering smell now that non-Validator code depends on it.
Move the general-purpose helpers (TrimAsciiWhitespace, ReplaceAllInPlace,
PathToUtf8, NormalizePathForComparison, JoinSortedSet) into a top-level
src/hdtStringUtils.h. The Validator header keeps its domain-specific helpers
(XSD/XPath/NIF/validation messages) and includes the general one, so its five
consumers compile unchanged.
